Moving XML cleanup logic into new XmlHandler class
